requirements.txt
alembic==0.9.9
Babel==2.7.0
blinker==1.4
click==6.7
dominate==2.3.1
Flask==0.12
Flask-BabelEx==0.9.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Principal==0.4.0
Flask-pymysql==0.2.3
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
passlib==1.7.1
PyJWT==1.6.4
PyMySQL==0.9.3
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2019.1
six==1.11.0
speaklater==1.3
SQLAlchemy==1.3.5
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2
